refactor(host): move ctl_nats back under wasmbus/nats
