+++ Cargo.toml
@@ -135,7 +135,6 @@ clap = { workspace = true, features = [
     "usage",
 ], optional = true }
 clap-markdown = { workspace = true, optional = true }
-futures = { workspace = true }
 nkeys = { workspace = true, optional = true }
 regex = { workspace = true, optional = true }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread", "signal"] }
